Split shell disconnect reconciliation tests
